+++ CHANGELOG.md
@@ -13,6 +13,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added mruby-env mrbgem.
 - Added mruby-logger mrbgem.
 
+### Changed
+
+- The name for the file used as entry is now "boot" instead of "entrypoint", I find
+  it more fitting and less verbose.
+
 ## [0.0.2] - 2023-11-09
 
 ### Added